requirements.txt
boto==2.48.0
boto3==1.7.46
botocore==1.10.46
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-pandas==0.5.1
django-rest-auth==0.9.3
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.8.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
numpy==1.14.5
pandas==0.23.1
Pillow==5.1.0
pipenv==2018.5.18
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
